fix(Dockerfile): correct paths to web content and db
The original version was populated with a 0 byte `server.sqlite` that is needed
to show who is allowed on the server and the `WebContent` folder wasn't copied
over. This ensures that both of those are fixed and for the first time allows
the application to execute properly.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>